Deleted view folder, moved all .vue files to components folder
